Add files via upload
I created a collection of different patterns to express the adjective "sparkle" and also used the keys on the keyboard to spell "sparkle" and have each pattern correspond to each letter. Some of the patterns are interactive. Check out the last pattern I did that is very similar to the confetti concept that I had talked about last time I pushed the project. I used a combination of different codes to create it and tweaked it from the code in Nature of Code chapter!